The Problem: FIXED
The issue that we were running into was that the Targets were never
cleared for a Cell.  If a Cell was previously occupied, then it would
still have a Person in its Targeted list and would result in a Person
to stop moving.  The Targeted list is now cleared whenever a person
moves to that Cell.